requirements.txt
Django>=6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
python-decouple>=3.8
requests>=2.31.0
django-cors-headers>=4.3.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
redis>=5.0.0
